Fixing gcc build of long long interface, fixing a horde of build warnings

+++ packages/pamgen/src/brick_inline_mesh_desc.C
@@ -340,7 +340,7 @@ void Brick_Inline_Mesh_Desc::Populate_Coords(double * coords,
 
 /****************************************************************************/
 {
-  for(long long gnv = 0;gnv < global_node_vector.size();gnv ++){
+  for(unsigned gnv = 0;gnv < global_node_vector.size();gnv ++){
     long long the_node = global_node_vector[gnv];
     long long global_k = the_node/knstride;
     long long global_j = (the_node-global_k*knstride)/jnstride;
